pyproject.toml
[project]
name = "epenguin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "flask-apscheduler>=1.13.1",
    "flask-login>=0.6.3",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "google-auth-oauthlib>=1.2.2",
    "markdown>=3.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
